Endpoints con la seguridad según roles y métodos a acceder

+++ MicroservicioAlquileres/trabajoPracticoIntegrador/src/main/java/com/bda/trabajoPracticoIntegrador/Controller/AlquileresController.java
@@ -1,9 +1,12 @@
 package com.bda.trabajoPracticoIntegrador.Controller;
 
+import com.bda.trabajoPracticoIntegrador.Dtos.EstacionDto;
 import com.bda.trabajoPracticoIntegrador.Entity.Alquileres;
 import com.bda.trabajoPracticoIntegrador.Service.Interface.AlquileresService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.HttpClientErrorException;
@@ -32,25 +35,77 @@ public class AlquileresController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Alquileres> getById(@PathVariable int id) {
+    public ResponseEntity<AlquilerDto> getById(@PathVariable int id) {
         Alquileres alquileres = service.getById(id);
         if(alquileres != null) {
-            return ResponseEntity.ok(alquileres);
+            int estacionRetiroId = alquileres.getEstacionRetiro();
+            int estacionDevolucionId = alquileres.getEstacionDevolucion();
+
+            EstacionDto estacionRetiroDto = getEstacionFromOtherService(estacionRetiroId);
+            EstacionDto estacionDevolucionDto = getEstacionFromOtherService(estacionDevolucionId);
+
+            AlquilerDto alquilerDto = new AlquilerDto();
+            alquilerDto.setId(alquileres.getId());
+            alquilerDto.setIdCliente(alquileres.getIdCliente());
+            alquilerDto.setEstado(alquileres.getEstado());
+            alquilerDto.setEstacionRetiro(estacionRetiroDto);
+            alquilerDto.setEstacionDevolucion(estacionDevolucionDto);
+            alquilerDto.setFechaHoraRetiro(alquileres.getFechaHoraRetiro());
+            alquilerDto.setFechaHoraDevolucion(alquileres.getFechaHoraDevolucion());
+            alquilerDto.setMonto(alquileres.getMonto());
+
+            return ResponseEntity.ok(alquilerDto);
+
         } else {
             return ResponseEntity.notFound().build();
         }
     }
 
-    @PostMapping
-    public Alquileres addAlquiler(@RequestBody Alquileres alquiler) {
-        
-
-        return service.add(alquiler);
+    private EstacionDto getEstacionFromOtherService(int estacionId) {
+        String url = "http://localhost:8084/api/estaciones/" + estacionId;
+        return restTemplate.getForObject(url, EstacionDto.class);
     }
 
-    @DeleteMapping("/delete/{id}")
+    @PostMapping
+    public ResponseEntity<AlquilerDto> addAlquiler(@RequestBody AlquilerDto alquiler) {
+        Alquileres alquileres = new Alquileres();
+
+        int estacionRetiroId = alquiler.getEstacionRetiro().getId();
+        int estacionDevolucionId = alquiler.getEstacionDevolucion().getId();
+        EstacionDto estacionRetiroDto = getEstacionFromOtherService(estacionRetiroId);
+        EstacionDto estacionDevolucionDto = getEstacionFromOtherService(estacionDevolucionId);
+
+        if (estacionRetiroId == 0 || estacionDevolucionId == 0) {
+            // Manejo de estaciones no válidas, por ejemplo, puedes devolver un error 400 (Bad Request)
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        alquileres.setIdCliente(alquiler.getIdCliente());
+        alquileres.setEstado(alquiler.getEstado());
+
+        alquiler.setEstacionRetiro(estacionRetiroDto);
+        alquiler.setEstacionDevolucion(estacionDevolucionDto);
+
+        Alquileres nuevoAlquiler = service.add(alquileres);
+
+        return ResponseEntity.ok(alquiler);
+    }
+
+
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
-        service.delete(id);
-        return ResponseEntity.noContent().build();
+        Alquileres alquileres = service.getById(id);
+
+        if (alquileres != null) {
+            int estacionRetiroId = alquileres.getEstacionRetiro();
+            int estacionDevolucionId = alquileres.getEstacionDevolucion();
+
+            // Realiza la lógica para eliminar la relación con estaciones si es necesario
+
+            service.delete(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
